requirements.txt
customtkinter>=5.2.2
tkinterdnd2>=0.3.0
opencv-python>=4.8.0
Pillow>=10.0.0
requests>=2.31.0
psutil>=5.9.0
numpy>=1.24.0
